requirements.txt
# ----------------------------------
# V2/V3 核心手写版 & 底层版 依赖
# ----------------------------------
# 核心计算库 (指定CPU版本以保证在无GPU环境下的兼容性)
torch==2.3.1
numpy==1.26.4

# 配置文件解析
pyyaml==6.0.1

# 数据处理
pandas==2.2.2

# 进度条
tqdm==4.66.4

wandb==0.17.0
swanlab==0.2.6


# 高性能正则表达式库 (用于分词器)
regex==2025.11.3
# ----------------------------------
# V1 快速原型版 依赖 (可选安装)
# ----------------------------------
# Hugging Face 生态
transformers==4.41.2
datasets==2.19.2
tokenizers==0.19.1
accelerate==0.30.1
peft==0.11.1
trl==0.9.4

# ----------------------------------
# 日志与监控 (推荐安装)
# ----------------------------------
wandb==0.17.0
tensorboard==2.16.2